TurboGears/requirements.txt
SQLAlchemy
TurboGears2
sprox
psycopg2
tw2.forms
mako